Addad sto passage to model, repository etc.

+++ ie-city-bus/src/main/java/bjdodo/ie_city_bus/repository/VehicleRepository.java
@@ -1,12 +1,12 @@
 package bjdodo.ie_city_bus.repository;
 
-import bjdodo.ie_city_bus.model.Vehicle;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+
+import bjdodo.ie_city_bus.model.Vehicle;
 
 public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
 	
 	@Query("select count(v) FROM Vehicle v WHERE v.duid = ?1")
-	int countByDuid(String duid);
+	long countByDuid(String duid);
 }
